Share code regarding receipt validation
